Add support for ~/.aws/config file
Fix for https://github.com/remind101/assume-role/issues/7 .  Tested with both
~/.aws/config and ~/.aws/roles.  Did NOT test that a ~/.aws/config profile that
doesn't have an MFA token works since it would have required setting up test
users on my side.  I did test that it can parse an entry that doesn't have an
MFA and sets the value in the map to nil.  If that is what the YAML parser does
it should work.

Prints a deprecation warning if running against ~/.aws/roles.

- Add gp-ini/ini dependency to vendors (https://github.com/go-ini/ini)
- Move old loadConfig to loadConfigFromRoleFile
- Add new loadConfigFromConfigFile function
- Make new loadConfig function that delegates to the other two
- Run go fmt